add authentication and error return if unauthorized
Former-commit-id: ac649ae6be96663cbef170a027b5579425b605cb

+++ app/controllers/playlists_controller.rb
@@ -10,7 +10,7 @@ class PlaylistsController < ApplicationController
 				@musics = @playlist.musics.ordered
 				render 'show'
 			else
-				redirect_to user_path(current_user), flash: {notice: 'Unauthorized access'}
+				redirect_to user_path(current_user), flash: {alert: 'Unauthorized access'}
 			end
 		else
 			redirect_to user_path(current_user), flash: {notice: 'Playlist Not Found'}
@@ -19,11 +19,15 @@ class PlaylistsController < ApplicationController
 
 	def edit
 		if @playlist = Playlist.find_by_id(params[:id])
-			@musics = @playlist.musics.ordered
-			@new_musics = Music.all_except(@musics).order(:title)
-			render 'edit'
+			if @playlist.user == current_user
+				@musics = @playlist.musics.ordered
+				@new_musics = Music.all_except(@musics).order(:title)
+				render 'edit'
+			else
+				redirect_to user_path(current_user), flash: {alert: 'Unauthorized Access'}
+			end
 		else
-			redirect_to user_path(current_user), flash: {alert: 'Unauthorized Access'}
+			redirect_to user_path(current_user), flash: {notice: 'Playlist Not Found'}
 		end
 	end
 
@@ -38,34 +42,48 @@ class PlaylistsController < ApplicationController
 		else
 			respond_to do |format|
 				format.html { redirect_to user_path(current_user), :flash => { :alert => @playlist.errors.full_messages} }
-				format.json { render json: @playlist.errors, status: 422 }
+				format.json { render json: @playlist.errors.full_messages, status: 422 }
 			end
 		end
 	end
 
 	def update
-		if @playlist = Playlist.find_by_id(params[:id])
-			if @playlist.update(playlist_params)
+		if @playlist = Playlist.find_by_id(params[:id]) 
+			if @playlist.user == current_user && @playlist.update(playlist_params)
 				respond_to do |format|
 					format.html { 
-						redirect_back(fallback_location: edit_playlist_path(params[:id]), flash: {notice: "Changes Saved"}) 
+						redirect_back(fallback_location: edit_playlist_path(@playlist), flash: {notice: "Changes Saved"}) 
 					}
 					format.json { render json: @playlist } 
 				end
 			else
-				redirect_to user_path(current_user), flash: {alert: @playlist.errors.full_messages}
+				respond_to do |format|
+					format.html { 
+						redirect_back(fallback_location: playlist_path(@playlist), flash: {alert: @playlist.errors.full_messages})
+					}
+					format.json { render json: @playlist.errors.full_messages, status: 422 }
+				end
 			end
 		else
-			redirect_to user_path(current_user)
+			respond_to do |format|
+				format.html { 
+					redirect_to user_path(current_user), flash: {notice: 'Playlist Not Found'} 
+				}
+				format.json { render json: 'Playlist Not Found', status: 404 } 
+			end
 		end
 	end
 
 	def destroy
-		if playlist = Playlist.find_by_id(params[:id])
-			playlist.destroy
-			redirect_to user_path(current_user), flash: {notice: "#{playlist.name} has been removed"}
+		if playlist = Playlist.find_by_id(params[:id]) 
+			if playlist.user == current_user
+				playlist.destroy
+				redirect_to user_path(current_user), flash: {notice: "#{playlist.name} has been removed"}
+			else
+				redirect_to user_path(current_user), flash: {alert: 'Unauthorized Access'}
+			end
 		else
-			redirect_to user_path(current_user), :flash => { :alert => "Unknown error occured"}
+			redirect_to user_path(current_user), flash: {notice: 'Playlist Not Found'}
 		end
 	end
